Update to use dynamic loading pattern
